moved and renamed notifyGroup function into notifications/sender.js
